renaming and removing files
