+++ printcetnrum_web.client/src/app/models/diploma-container.model.ts
