+++ printcetnrum_web.client/src/app/models/stamp-container.model.ts
