+++ printcetnrum_web.client/src/app/pages/diploma-page/diploma-page.component.ts
@@ -2,7 +2,7 @@ import { Component, ElementRef, HostListener, ViewChild, inject } from '@angular
 import { ActivatedRoute } from '@angular/router';
 import { SnackBarUtil } from '../../shared/snackbar-util';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { DiplomaTextBox } from '../../models/diploma';
+import { DiplomaTextBox } from '../../models/diploma-container.model';
 import {DesignFilesHandlerService} from "../../services/design-files-handler.service";
 
 @Component({
