--- printcetnrum_web.client/src/app/pages/edit-photo-page/edit-photo-page.component.css
@@ -1,66 +1,0 @@
-.displayed-image {
-  width: 100%; 
-  max-width: 500px; 
-  height: auto; 
-  border: 2px solid #ccc; 
-  background-color: #f0f0f0; 
-}
-
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-  background-color: #f8f9fa;
-}
-
-.image-container {
-  text-align: center;
-}
-
-.image-frame {
-  position: relative;
-  width: 80%; 
-  height: 80%; 
-  margin: 10px auto;
-  margin-top: 1px; 
-  border: 2px solid #ccc;
-  overflow: hidden; 
-  background-color: #f0f0f0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-
-img {
-  max-width: 100%; 
-  max-height: 600px;
-  object-fit: contain;
-}
-
-.crop-controls {
-  margin-top: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-  .crop-controls input {
-    margin-left: 10px;
-    margin-right: 10px;
-  }
-
-  .crop-controls button {
-    margin-left: 10px;
-  }
-
-button {
-  margin: 5px; 
-  padding: 8px 16px;
-  font-size: 14px; 
-}
-
-canvas {
-  display: none;
-}
