--- printcetnrum_web.client/src/app/pages/edit-photo-page/edit-photo-page.component.html
@@ -1,31 +1,0 @@
-<body>
-  <div class="container">
-    <div class="image-container">
-      <input type="range" min="0" max="200" [(ngModel)]="brightness" (input)="updateBrightness()" />
-      <button class="btn btn-primary" (click)="revertChanges()">Revert</button>
-      <button class="btn btn-primary" (click)="saveChanges()">Save Changes</button>
-      <button class="btn btn-primary" (click)="download()">Download Image</button>
-      <p>{{ this.brightness }}</p>
-
-      <div class="image-frame">
-        <img [style.filter]="brightnessStyle"
-             [src]="imageUrl"
-             alt="Image"
-             id="image"
-             crossorigin="anonymous"
-             (load)="onImageLoad()" />
-      </div>
-
-      <div class="crop-controls">
-        <label for="cropWidth">Width:</label>
-        <input id="cropWidth" type="number" [(ngModel)]="cropWidth" min="50" />
-        <label for="cropHeight">Height:</label>
-        <input id="cropHeight" type="number" [(ngModel)]="cropHeight" min="50" />
-        <button class="btn btn-secondary" (click)="applyCrop()">Apply Crop</button>
-      </div>
-
-      <canvas id="canvas" style="display: none;"></canvas>
-      <p>{{ this.imageUrl }}</p>
-    </div>
-  </div>
-</body>
