--- printcetnrum_web.client/src/app/pages/edit-photo-page/edit-photo-page.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { EditPhotoPageComponent } from './edit-photo-page.component';
-
-describe('EditPhotoPageComponent', () => {
-  let component: EditPhotoPageComponent;
-  let fixture: ComponentFixture<EditPhotoPageComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [EditPhotoPageComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(EditPhotoPageComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
