+++ printcetnrum_web.client/src/app/shared/file-card-component/file-card-component.component.html
@@ -24,7 +24,6 @@
           <i class="bi bi-printer"></i>
           {{ file.shouldPrint ? 'Unmark' : 'Mark for Print' }}
         </button>
-        <button *ngIf="isImage(file.extension)" class="btn btn-secondary" (click)="editFile.emit(file.id)">Edit</button>
         <button *ngIf="file.isStamp" class="btn btn-secondary" [routerLink]="['/editStamp', file.id]">Edit Stamp</button>
         <button *ngIf="file.isDiploma" class="btn btn-secondary" [routerLink]="['/editDiploma', file.id]">Edit Diploma</button>
         <button class="btn btn-danger" (click)="openDeleteModal.emit(file.id)">Remove</button>
